extract the scrolling and focus logic in separate file
